+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # CHANGELOG
 
 ## 0.21.0
@@ -14,6 +15,7 @@
 * Support passing metrics configuration as an external ConfigMap
 * Enable CORS configuration for Cruise Control
 * Add support for rolling individual Kafka or ZooKeeper pods through the Cluster Operator using an annotation
+* Add support for Topology Spread Constraints in Pod templates
 
 ### Deprecations and removals
 * The `metrics` field in the Strimzi custom resources has been deprecated and will be removed in the future. For configuring metrics, use the new `metricsConfig` field and pass the configuration via ConfigMap.